댓글 존재 시 게시물 삭제 불가 기능 추가

+++ src/main/java/com/yyi/projectStudy/controller/ProjectCommentController.java
@@ -54,7 +54,7 @@ public class ProjectCommentController {
 
     /* 게시글의 댓글 수 조회 */
     @GetMapping("/count/{id}")
-    public @ResponseBody Long commentCount(@PathVariable("id") Long id) {
+    public @ResponseBody int commentCount(@PathVariable("id") Long id) {
         return projectCommentService.count(id);
     }
